// src/api/generated/models/bookInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * API каталога книг
 * REST API для управления каталогом книг и авторов.
 * Поддерживает роли: гость (просмотр) и пользователь (CRUD).
 * Аутентификация через Bearer-токен (JWT).
 *
 * OpenAPI spec version: 1.0.0
 */

export interface BookInput {
  title?: string;
  year?: number;
  description?: string;
  isbn?: string;
  author_ids?: number[];
}